Organized the files, began implementing database in userManager
User manager will store the current user data that is not tracked by FirebaseAuth in a private static user field that will be publically accessable. User class needs to be updated to be in line with the current constants and needs to be updated to implement the constructor that takes in a hashmap that will contain the user data stored in the users collection

ImageEncodingUtility has been added and will be used to encode and decode images like profile pictures and location pictures.

I organized classes into packages to make managing them easier, sorry if this causes problems when combining merges!