Implemented database functions into UserManager signin method, added FirestoreUtility class
FirestoreUtility needs to be fully realized, we need to find a way to fully construct a schedule by using one or multiple database calls. if its multiple the construction needs to happen piecemeal style by adding them incrementally to the user object. Im getting kinda burnt out on this project so im going to take a break and do something else, might begin work on making the locations XML activity